Possible solution to the question
I've used a vertical flexbox to seperate the title and the plants.
Next, I've used a row flexbox to seperate the plants. Then, each
plant I made a column flexbox with 200px width and everything
is centered using align-itmes or text-align. Used gaps too. Not
sure if optimal solution, but I think it works.